requirements.txt
boto3>=1.28.0
PyYAML>=6.0
botocore>=1.31.0
PyJWT>=2.6.0
requests>=2.28.0
cryptography>=41.0.0